+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.13.7] - 2025-10-01
 
+### Added
+- Extra validation for authn-iam configuration. CNJR-10225
+
 ## [0.13.6] - 2025-09-30
 
 ### Added